requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2     
attrs==22.2.0
certifi==2022.12.7       
charset-normalizer==2.1.1
comtypes==1.1.14
frozenlist==1.3.3        
greenlet==2.0.2
idna==3.4
iso8601==1.1.0
multidict==6.0.4
obs-websocket-py==0.5.3
pypiwin32==223
pyttsx3==2.90
pywin32==305
redis==4.4.2
requests==2.28.2
six==1.16.0
spotipy==2.22.1
SQLAlchemy==2.0.1
twitchio==2.5.0
typing_extensions==4.4.0
urllib3==1.26.14
websocket-client==1.5.0
yarl==1.8.2